add include, it's work, but without defense

+++ srcs/server.c
@@ -6,46 +6,54 @@
 /*   By: pantigon <pantigon@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/21 19:45:12 by pantigon          #+#    #+#             */
-/*   Updated: 2021/09/22 18:57:41 by pantigon         ###   ########.fr       */
+/*   Updated: 2021/09/23 18:57:41 by pantigon         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <unistd.h>
-#include <stdio.h>
-#include <signal.h>
-#include "libft.h"
+#include "minitalk.h"
+
+t_data	g_data;
+
+void	init_data(void)
+{
+	g_data.byte = 0;
+	g_data.chr = 0;
+}
+
+void	client_id(pid_t cid)
+{
+	g_data.cid = cid;
+	ft_putstr_fd("client [", 1);
+	ft_putnbr_fd(cid, 1);
+	ft_putstr_fd("]: ", 1);
+}
 
 void	change_signal(int signal, siginfo_t *siginfo, void *context)
 {
-	if (signal == SIGUSR1)
-		ft_putstr_fd("1\n", 1);
-	else if (signal == SIGUSR2)
-		ft_putstr_fd("0\n", 1);
 	(void)context;
-	kill(siginfo->si_pid, SIGUSR1);
-	// printf("OK\n");
-	// char	c;
-	// int		i;
-
-	// i = 8;
-	// while (i > 0)
-	// {
-	// 	if (signal == SIGUSR1)
-	// 	{
-
-	// 	}
-	// 	else
-	// 	{
-			
-	// 	}
-	// 	--i;
-	// }
+	if (g_data.cid != siginfo->si_pid)
+		client_id(siginfo->si_pid);
+	if (g_data.byte == 8)
+	{
+		ft_putchar_fd(g_data.chr, 1);
+		if (g_data.chr == 0)
+		{
+			ft_putchar_fd('\n', 1);
+			kill(siginfo->si_pid, SIGUSR1);
+		}
+		init_data();
+		return ;
+	}
+	if (signal == SIGUSR1)
+		g_data.chr |= (1 << g_data.byte);
+	g_data.byte++;
 }
 
 int	main(void)
 {
-	struct sigaction sig;
+	struct sigaction	sig;
 
+	init_data();
 	ft_putstr_fd("server id: ", 1);
 	ft_putnbr_fd(getpid(), 1);
 	ft_putstr_fd("\n", 1);
